requirements.txt
python-dotenv
flask
flask_sqlalchemy
flask_smorest
flask_jwt_extended
passlib
flask_jwt_extended
redis
celery